+++ backend/ttabong/src/main/java/com/ttabong/dto/sns/response/RecruitReviewByVolResponseDto.java
@@ -16,8 +16,8 @@ import java.util.List;
 @Builder
 public class RecruitReviewByVolResponseDto {
     // 후기 정보
-    private Long reviewId;
-    private Long recruitId;
+    private Integer reviewId;
+    private Integer recruitId;
     private String title;
     private String content;
     private Boolean isDeleted;
@@ -25,15 +25,15 @@ public class RecruitReviewByVolResponseDto {
     private LocalDateTime createdAt;
 
     // 작성자 정보
-    private Long writerId;
+    private Integer writerId;
     private String writerName;
 
     // 그룹 정보
-    private Long groupId;
+    private Integer groupId;
     private String groupName;
 
     // 기관 정보
-    private Long orgId;
+    private Integer orgId;
     private String orgName;
 
     // 이미지 리스트
